Implementing Rate Limiting Middleware
- In order to provide a simple security
measure against Denial of Service
attacks, a middleware for rate limiting
was implemented. It can be configured
via application.json for time window
and max requests.

- The middleware is optional and will be
active only if the configurations
are provided in the application.json file.

- It is a simple middleware for single
instances applications. In the case
of distributed systems and containerized
environments it's recommended to implement
rate limiting in the Load Balancer layer.