Development (#43)
* Improving caching strategy

- Add a new configuration option in the cache section in
  the application.json file, called ignore_headers which
  will accept an array of header names that should not be
  considered when determining cache usage

* Add HTTPS/SSL support using OpenSSL

- This commit introduces HTTPS/SSL support
to the framework, enabling developers to
secure their applications with encrypted
communication. The implementation leverages
the Ruby OpenSSL gem for handling SSL
certificates and keys.

- A new method in the Server class to set up
SSL context, which wraps the existing TCP
server with HTTPS security.

- The SSL certificate and private key file
paths are specified in the application.json
file under the macaw configuration section.

- SSL support is enabled only when the ssl
field is present in the macaw config section
of the application.json file.

* Fixing a bug of noMethodError in cache aspect

- When the ignore_headers configuration where not
provided in the application.json while cache was active,
this generated an error of noMethodError when trying to
process cache strategy. Null safety operators on the
array search corrected the problem.

* Fixing a bug of noMethodError in cache aspect

- When the ignore_headers configuration where not
provided in the application.json while cache was active,
this generated an error of noMethodError when trying to
process cache strategy. Null safety operators on the
array search corrected the problem.

* Fixing cache ignored headers documentation

- In the documentation and in the code, the ignored_headers
array was expected inside the rate limiting configuration,
which was a design mistake. This commit corrects it moving
the configuration to the cache configuration.

* Create SECURITY.md (#41)

* Implementing session feture (#42)

- With this commit is introduced a new session feature,
which permits to store client specific data server side
in memory for individual clients based on their IP.