Potentially fixing the thread starvation issue (#65)
- In this commit were identified and possibly fixed
the issue of thread starvation. When the client
close connection abruptly, an error were being raised
in the ensure clause that closed the connection, since
it was already terminated. This was fixed and, to
prevent other sources of thread starvation, a mechanism
to maintain the threads in the pool was created.