Fix/commit history (#69)
* v1.0.5 (#66)

* Potentially fixing the thread starvation issue (#65)

- In this commit were identified and possibly fixed
the issue of thread starvation. When the client
close connection abruptly, an error were being raised
in the ensure clause that closed the connection, since
it was already terminated. This was fixed and, to
prevent other sources of thread starvation, a mechanism
to maintain the threads in the pool was created.

* Creating log sanitization to prevent log forging (#47)

* Creating log sanitization to prevent log forging

- In order to filter logs to prevent log forging
attacks and to make them more legible, the
module LogDataFilter was created. It will sanitize
every log registered by the logging aspect.

- Also, more tests were created for Macaw class
and Request Data Filter module.

* Change minimum Ruby version from 2.7 to 3.0.0

- Since support for 2.7 expired now in 2023,
Macaw Framework will be developed focusing
on versions > 3.0.0

* Revert "Creating log sanitization to prevent log forging (#47)" (#48)

This reverts commit f133737b96c5b30277b73e361d3a90d6586798ee.

* Creating log sanitization to prevent log forging (#59)

- In order to filter logs to prevent log forging
attacks and to make them more legible, the
module LogDataFilter was created. It will sanitize
every log registered by the logging aspect.

- Also, more tests were created for Macaw class
and Request Data Filter module.

* Implementing support for other SSl keys than RSA (#68)

- In order to provide more flexibility for
developers, support for Elliptic Curve
Digital Signature Algorithm (ECDSA) was
introduced.