1.1.0 (#74)
* Potentially fixing the thread starvation issue (#65)

- In this commit were identified and possibly fixed
the issue of thread starvation. When the client
close connection abruptly, an error were being raised
in the ensure clause that closed the connection, since
it was already terminated. This was fixed and, to
prevent other sources of thread starvation, a mechanism
to maintain the threads in the pool was created.

* Implementing support for other SSl keys than RSA (#68)

- In order to provide more flexibility for
developers, support for Elliptic Curve
Digital Signature Algorithm (ECDSA) was
introduced.

* Fix/commit history (#69)

* v1.0.5 (#66)

* Potentially fixing the thread starvation issue (#65)

- In this commit were identified and possibly fixed
the issue of thread starvation. When the client
close connection abruptly, an error were being raised
in the ensure clause that closed the connection, since
it was already terminated. This was fixed and, to
prevent other sources of thread starvation, a mechanism
to maintain the threads in the pool was created.

* Implementing support for other SSl keys than RSA (#68)

- In order to provide more flexibility for
developers, support for Elliptic Curve
Digital Signature Algorithm (ECDSA) was
introduced.

* Improving log readability (#72)

- To improve log readability, the automatic
logs now exclude / characters.

- To improve performance and customization,
the Log Aspect is now optional. Ite can be
turned off setting custom_log: nil when
declaring a new instance of Macaw

- Including a CONTRIBUTING.md

* Properly handling server shutdown (#73)

- In order to wait for running threads to
finish their tasks before shutting down
and not accept more requests in the meantime,
a shutdown method was included in the server
along with modifications to the method that
maintain threads.