Fixing bug of framework not starting on non-Unix machines (#98)
- Since the new feature `public folder` relies on a
system-dependent feature of Ruby, some systems were
not able to run applications with MacawFramework. To fix
this, we set the dir: property default to nil and
skipped the use of this function when the value of
dir is nil. This way, users should be able to use
the public folder feature only if they want by setting
manually the value of dir at instantiation.