Fixing log message on declaration of endpoints (#107)
Before this fix, when you declare a new endpoint sinalizing a /
as the first character, the log message would print // at the start
of the name of the endpoint. Now this behavior is fixed. Now
the logs will always print the endpoint name with a single /
independent if it was declared with / or not.